refactor: move dto package into model
